Factor ServerSetSource into a parent class of AuroraProxySource.
