+++ .gitignore
@@ -15,3 +15,4 @@ DS_Store
 dist/*
 *.egg-info/
 *.iml
+*~
