+++ tellapart/aurproxy/source/__init__.py
@@ -17,3 +17,4 @@ from .source import ProxySource
 from .sources.api import ApiSource
 from .sources.aurora import AuroraProxySource
 from .sources.static import StaticProxySource
+from .sources.serverset import ServerSetSource
